Added useState import and eventHandling to Form component
-Imported { useState }
-Created state variable name
-Event Handling for form input changing, sets name variable to form value on change
-Event Handling for submitting, uses prop.addTask with the form value, then resets value to empty string

App.js sends in a function as a prop to the Form component. This function takes an entered String and adds a new task to the tasks array. This updates the DOM and displays a new task.
This was done so that the Form component does not need to do any creation of elements, rather it keeps track of its data and uses its parent function addTask to update is parents state.
This well displays the functionality of hooks and states in React.