toggleTaskCompleted and deleteTask
-toggleTaskCompled changes the Apps local memory to be in line with the browser caches memory of the checkboxes "checked" state
-deleteTask removes task form tasks array, which updates the displayed taskList

These are sent into the Todo component for use in event handling their respective button clicks. Without the toggleTaskCompleted, the apps memory would not align with the browsers knowledge of the checkboxes state.
Deleting a task must remove it from the taskList, so we take a callback function in as a prop to the Todo component, ask when the button is clicked, and update the state of the App to reflect that the user has asked to remove a task form the display.